requirements.txt
Selenium
Pytest
Pillow
pytest-html